refactor(client): decode source inputs outside CLI
