+++ crates/coral-client/src/lib.rs
@@ -5,6 +5,7 @@
 //! - endpoint dialing into the generated gRPC transport surface
 //! - lightweight shared Arrow IPC decoding helpers
 //! - lightweight shared result-format rendering used by CLI and MCP
+//! - lightweight client-side decoding helpers for shared transport DTOs
 //!
 //! It does **not** currently try to present a richer domain SDK. Callers that
 //! need more abstraction should add it above this crate rather than widening
@@ -19,6 +20,7 @@ mod error;
 mod grpc;
 pub mod local;
 mod propagation;
+mod sources;
 mod status_error;
 
 use std::io::Cursor;
@@ -36,6 +38,7 @@ pub use client::{
     default_workspace,
 };
 pub use error::{ClientError, QueryResultError};
+pub use sources::{SourceInputDecodeError, manifest_input_from_proto};
 pub use status_error::{
     CORAL_ERROR_DOMAIN, CoralQueryError, DecodedStatusError, decode_status_error,
 };
