+++ docs/reference/source-spec-reference.mdx
@@ -315,8 +315,9 @@ In this minimal example:
 
 - `response: {}` means "use the default response rules" for an HTTP table
 - with the default response rules, Coral treats the selected response value directly as rows unless you configure a different `row_strategy`
-- `inputs` declares the install-time variables and secrets for this source in stable authored order. At install time, `coral source add` reads each key from an environment variable of the same name (or prompts for them when run with `--interactive`). At runtime, installed inputs are surfaced through `coral.inputs`.
-- `{{input.API_TOKEN}}` is resolved from the store implied by the declared input kind
+- `inputs` declares values Coral collects and stores when the source is added
+- `auth` declares how stored values are used to authenticate outgoing HTTP requests
+- `{{input.API_TOKEN}}` is resolved from the secret store because `API_TOKEN` is declared as `kind: secret`
 
 ### HTTP table functions
 
@@ -378,15 +379,13 @@ argument. Function request values can then reference those bound arguments with
 
 ### HTTP authentication
 
-The top-level `auth` block declares how Coral authenticates outgoing requests. Pick one of three `type` values:
+The top-level `auth` block declares how Coral authenticates outgoing requests. It does not collect or store credentials; reference values declared under [Source inputs](/reference/source-spec-reference#source-inputs) when auth needs API keys, bearer tokens, usernames, passwords, or signing material. Pick one of three `type` values:
 
 | `type` | Purpose |
 | ------ | ------- |
 | `HeaderAuth` | Attach one or more declarative auth headers (Bearer tokens, API keys, pre-signed headers). |
 | `BasicAuth` | HTTP Basic. Coral base64-encodes `username:password` into `Authorization: Basic ...`. |
 | `CustomAuth` | Use a named authenticator for auth schemes that need request signing, such as AWS SigV4. |
-
-Credentials for authentication are declared in [Source inputs](/reference/source-spec-reference#source-inputs).
 
 #### `HeaderAuth`
 
@@ -671,7 +670,9 @@ pagination:
 
 ## Source inputs
 
-HTTP source specs can declare the variables and secrets this source needs under a top-level `inputs` map. At install time, `coral source add` collects each input from an environment variable matching the key, or prompts interactively when you pass `--interactive`.
+HTTP source specs can declare the variables and secrets this source needs under a top-level `inputs` map. Inputs are where credentials and other configuration values are set up. They are not where those values are sent. Runtime request fields, such as `auth`, `request_headers`, table request headers, query params, and body fields, decide where stored inputs are used.
+
+At install time, `coral source add` collects each input from an environment variable matching the key, or prompts interactively when you pass `--interactive`.
 
 - **Variables**: non-secret configuration like base URLs or organization IDs
 - **Secrets**: API keys or bearer tokens
@@ -703,13 +704,15 @@ Rules:
 - `hint` is optional and shown alongside the input during `coral source add --interactive`
 - references elsewhere in the manifest use `{{input.KEY}}` templates or `from: input`
 - credential-like inputs such as API keys, tokens, passwords, secrets, private keys, and bearer or authorization values must be `kind: secret`; Coral rejects those names as variables because variable values are visible through source APIs and `coral.inputs`
+- declaring an input does not add it to any request by itself; reference it from `auth`, headers, query params, body fields, or other runtime request configuration where it should be used
 
 Secret inputs can also declare `credential.methods` metadata for install-time
-credential retrieval choices. Coral surfaces this metadata through source
-discovery and source-info responses, but the CLI still collects secret values
-through the normal environment variable or prompt path. This does not replace
-the top-level `auth` block: `credential.methods` describes install-time
-credential retrieval options, while `auth` is runtime request authentication.
+credential retrieval choices. When `credential` is omitted, Coral collects the
+secret the same way as `type: source_config`: from an environment variable by
+default, or from an interactive prompt when `coral source add --interactive` is
+used. `credential.methods` only controls how the secret is collected; runtime
+request authentication is still defined by `auth` or request headers that
+reference the input.
 
 ```yaml
 inputs:
@@ -753,7 +756,9 @@ fixed loopback HTTP callback URI:
 own OAuth client ID for a future OAuth retrieval flow, they must configure that
 exact redirect URI in the OAuth app.
 
-`pkce` is either `required` or `disabled`. Public clients can declare
+`flow.type` and `pkce` are required. Unspecified protobuf enum values are
+invalid and are rejected rather than treated as defaults. `pkce` is either
+`required` or `disabled`. Public clients can declare
 `client.id.default`, `client.id.input`, or both; an interactive `input` value
 overrides `default`. Confidential clients add `client.secret.input` and a
 required `transport` of `basic_auth` or `request_body`. Client secrets are sent
